some size_t replaced by unsigned

+++ src/base/iowrapper.h
@@ -38,7 +38,7 @@ public:
     Inputter(unsigned d, FILE* f, const char* path=nullptr) : FileWrapper(f, path) { reset(); vecsize_=d; }
     
     /// constructor which opens a file
-    Inputter(size_t d, const char* name, bool bin) : FileWrapper(name, bin?"rb":"r") { reset(); vecsize_=d; }
+    Inputter(unsigned d, const char* name, bool bin) : FileWrapper(name, bin?"rb":"r") { reset(); vecsize_=d; }
 
     /// return dimensionnally of vectors
     unsigned vectorSize() const { return vecsize_; }
